fix/meta-election: decide if others already did
Most of the logic for picking up a decision from other nodes is already
in-place:
In `MetaVoteCounts::new`, we pick any decision from any other voter and
apply it to ourselves.
In `MetaVotes::calculate_new_decision`, we use this decision from the
counts as the default value for a meta_vote.

Where the logic stopped working was in the multi-stage calculation of
the estimates:
* We first evaluate a temporary value for the meta-election that
  doesn't take into account the coin toss
* We then evaluate the coin toss value
* We then finalise the meta-election evaluation using the coin toss

Prior to this change, we wouldn't pass the other meta-votes when
finalising the meta-election evaluation. This would mean that
MetaVoteCounts::new wouldn't actually pick up on the decision that other
voters made anymore. This would mean that the decision reached in our
temp meta-vote would actually be discarded and we wouldn't shortcut the
decision process.

This bug caused issues in soak testings of the routing crate with the
real Parsec.